+++ Criollo/Apps/CriolloiOSApp/MultiRouteViewController.swift
@@ -10,7 +10,7 @@ import Criollo
 
 class MultiRouteViewController: CRViewController {
 
-    override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: NSBundle?, prefix: String?) {
+    override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: Bundle?, prefix: String?) {
         super.init(nibName: nibNameOrNil, bundle: nibBundleOrNil, prefix: prefix)
 
 
@@ -24,29 +24,29 @@ class MultiRouteViewController: CRViewController {
             response.send(NSStringFromCRHTTPMethod(request.method))
         }
 
-        self.add("/hello-c", viewController:HelloWorldViewController.self, withNibName: String(HelloWorldViewController.self), bundle: nil)
+        self.add("/hello-c", viewController:HelloWorldViewController.self, withNibName: String(describing: HelloWorldViewController.self), bundle: nil)
         self.add("/api", controller:APIController.self)
 
         // Public folders path
-        self.mount("/pub", directoryAtPath: "~", options:CRStaticDirectoryServingOptions.AutoIndex)
+        self.mount("/pub", directoryAtPath: "~", options:CRStaticDirectoryServingOptions.autoIndex)
 
         // Static file
-        self.mount("/file.txt", fileAtPath:"/etc/hosts", options:CRStaticFileServingOptions.Cache, fileName:"hosts", contentType:"text/plain", contentDisposition:CRStaticFileContentDisposition.Inline)
+        self.mount("/file.txt", fileAtPath:"/etc/hosts", options:CRStaticFileServingOptions.cache, fileName:"hosts", contentType:"text/plain", contentDisposition:CRStaticFileContentDisposition.inline)
     }
 
-    override func presentViewControllerWithRequest(request: CRRequest, response: CRResponse) -> String {
-        self.vars["title"] = String(self.dynamicType)
+    override func present(with request: CRRequest, response: CRResponse) -> String {
+        self.vars["title"] = String(describing: type(of: self))
 
         var text:String = String()
-        let env:NSDictionary! = request.valueForKey("env") as! NSDictionary
+        let env:NSDictionary! = request.value(forKey: "env") as! NSDictionary
         text += "<h3>Request Environment:</h2><pre>"
-        env.enumerateKeysAndObjectsUsingBlock({ (key,  object, stop) -> Void in
+        env.enumerateKeysAndObjects({ (key,  object, stop) -> Void in
             text += "\(key): \(object)\n"
         })
         text += "</pre>"
         self.vars["text"] = text
         
-        return super.presentViewControllerWithRequest(request, response: response)
+        return super.present(with: request, response: response)
     }
 
 }