+++ Criollo/Source/CRRequest.m
@@ -14,21 +14,25 @@
 #import "CRServer.h"
 #import "CRServer_Internal.h"
 #import "CRRequestRange.h"
+#import "CRRequestRange_Internal.h"
+#import "CRUploadedFile.h"
+#import "CRUploadedFile_Internal.h"
+#import "CRApplication.h"
 
 @implementation CRRequest {
     __strong NSMutableDictionary* _env;
 
-//    __block NSString * _multipartBoundary;
-//    __block dispatch_once_t _multipartBoundaryOnceToken;
-//
-//    __block NSString * _multipartBoundaryPrefixedString;
-//    __block dispatch_once_t _multipartBoundaryPrefixedStringOnceToken;
-//
-//    __block NSData * _multipartBoundaryPrefixedData;
-//    __block dispatch_once_t _multipartBoundaryPrefixedDataOnceToken;
-//
-//    NSString* currentMultipartBodyKey;
-//    NSString* currentMultipartFileKey;
+    __block NSString * _multipartBoundary;
+    __block dispatch_once_t _multipartBoundaryOnceToken;
+
+    __block NSString * _multipartBoundaryPrefixedString;
+    __block dispatch_once_t _multipartBoundaryPrefixedStringOnceToken;
+
+    __block NSData * _multipartBoundaryPrefixedData;
+    __block dispatch_once_t _multipartBoundaryPrefixedDataOnceToken;
+
+    NSString* currentMultipartBodyKey;
+    NSString* currentMultipartFileKey;
 }
 
 - (instancetype)init {
@@ -121,6 +125,25 @@
     [((NSMutableDictionary *)_query) setObject:obj forKey:key];
 }
 
+- (NSString *)description {
+    return [NSString stringWithFormat:@"%@ %@ %@", NSStringFromCRHTTPMethod(self.method), self.URL.path, NSStringFromCRHTTPVersion(self.version)];
+}
+
+- (BOOL)shouldCloseConnection {
+    BOOL shouldClose = NO;
+
+    NSString *connectionHeader = [self valueForHTTPHeaderField:@"Connection"];
+    if ( connectionHeader != nil ) {
+        shouldClose = [connectionHeader caseInsensitiveCompare:@"close"] == NSOrderedSame;
+    } else {
+        shouldClose = self.version == CRHTTPVersion1_0;
+    }
+
+    return shouldClose;
+}
+
+
+#pragma mark - Body parsing
 
 - (BOOL)parseJSONBodyData:(NSError *__autoreleasing  _Nullable *)error {
     BOOL result = NO;
@@ -142,101 +165,147 @@
 - (BOOL)parseMultipartBodyDataChunk:(NSData *)data error:(NSError *__autoreleasing  _Nullable * _Nullable)error {
     BOOL result = YES;
 
-//    NSLog(@"%s %lu bytes", __PRETTY_FUNCTION__, data.length);
-//    NSLog(@"%@", [[NSString alloc] initWithBytesNoCopy:(void*)data.bytes length:data.length encoding:NSUTF8StringEncoding freeWhenDone:NO]);
-//
-//    // Search for a boundary
-//    NSRange searchRange = NSMakeRange(0, data.length);
-//    NSRange nextBoundaryRange = [data rangeOfData:self.multipartBoundaryPrefixedData options:0 range:searchRange];
-//
-//    if ( nextBoundaryRange.location != NSNotFound ) {                                   // We have a boundary
-//
-//        // Check if we have something before the boundary
-//        if ( nextBoundaryRange.location != 0 ) {                                        // There is an existing chunk
-//
-//            // Extract the piece
-//            NSData* preambleData = [NSData dataWithBytesNoCopy:(void *)data.bytes length:nextBoundaryRange.location freeWhenDone:NO];
-//
-//            // Check if we have something buffered
-//            if ( self.bufferedBodyData.length > 0 ) {                                   // This is part of a field that
-//
-//                // Prepend the the buffered data to the one piece
-//                NSMutableData* bufferedAndPreambleData = [NSMutableData dataWithCapacity:(self.bufferedBodyData.length + nextBoundaryRange.location)];
-//                [bufferedAndPreambleData appendData:self.bufferedBodyData];
-//                [bufferedAndPreambleData appendData:preambleData];
-//
-//                // Flush the buffered data
-//                self.bufferedBodyData = nil;
-//
-//                // Call this method again with the combined data
-//                result = [self parseMultipartBodyDataChunk:bufferedAndPreambleData error:error];
-//            } else {
-//
-//                // Append the piece to the target if there is one otherwise discard it
-//                // (RFC 1341 says to discard anything before the first --boundary (the preamble)
-//                // http://www.w3.org/Protocols/rfc1341/7_2_Multipart.html
-//
-//                if (currentMultipartBodyKey != nil && self.body != nil) {
-//                    result = [self appendBodyData:preambleData forKey:currentMultipartBodyKey];
-//                } else if ( currentMultipartFileKey == nil && self.files != nil ) {
-//                    result = [self appendFileData:preambleData forKey:currentMultipartFileKey];
-//                }
-//
-//            }
-//
-//            if ( result ) {
-//                // Extract the remaining data
-//                NSData* nextChunkData = [NSData dataWithBytesNoCopy:(void *)data.bytes + nextBoundaryRange.location length:data.length - nextBoundaryRange.location freeWhenDone:NO];
-//
-//                // Call this method again with the remaining data
-//                result = [self parseMultipartBodyDataChunk:nextChunkData error:error];
-//            }
-//
-//        } else {                                                                        // We are starting with a new field
-//
-//            NSData* CRLFData = [CRConnection CRLFData];
-//            NSData* CRLFCRLFData = [CRConnection CRLFCRLFData];
-//
-//            // Read the header (starts after the --boundary and a CRLF and ends with CRLFCRLF)
-//            NSUInteger headerStartLocation = nextBoundaryRange.location + nextBoundaryRange.length + CRLFData.length;
-//            NSRange headerSearchRange = NSMakeRange(headerStartLocation, data.length - headerStartLocation);
-//
-//            NSRange headerTerminatorRange = [data rangeOfData:CRLFCRLFData options:0 range:headerSearchRange];
-//
-//            if ( headerTerminatorRange.location != NSNotFound ) {                                 // We have a header - all good
-//
-//                NSData* headerData = [NSData dataWithBytesNoCopy:(void *)data.bytes + headerSearchRange.location length:headerTerminatorRange.location - headerSearchRange.location freeWhenDone:NO];
-//                NSLog(@" * Header:\n%@", [[NSString alloc] initWithBytesNoCopy:(void *)headerData.bytes length:headerData.length encoding:NSUTF8StringEncoding freeWhenDone:NO] );
-//
-//                // Parse the multipart header
-////                NSMutableDictionary* parsedHeaderFields = [NSMutableDictionary alloc]
-////                NSString* headerString = [[NSString alloc] initWithBytesNoCopy:(void *)headerData length:headerData.length encoding:NSUTF8StringEncoding freeWhenDone:NO];
-////                NSArray<NSString *>* headerLines = [headerString componentsSeparatedByCharactersInSet:[NSCharacterSet newlineCharacterSet]];
-////                [headerLines enumerateObjectsUsingBlock:^(NSString * _Nonnull obj, NSUInteger idx, BOOL * _Nonnull stop) {
-////
-////                }];
-//            } else {                                                                    // There is no header something is very wrong
-//                *error = [NSError errorWithDomain:CRRequestErrorDomain code:CRRequestErrorMalformedBody userInfo:@{NSLocalizedDescriptionKey: NSLocalizedString(@"Unable to parse multipart data.",)}];
-//                return NO;
-//            }
-//
-//        }
-//
-//    } else {                                                                            // This is just a chunk of something
-//
-//        // Append the data to the target if there is one otherwise discard it
-//        // (RFC 1341 says to discard anything before the first --boundary (the preamble)
-//        // http://www.w3.org/Protocols/rfc1341/7_2_Multipart.html
-//
-//        if (currentMultipartBodyKey != nil && self.body != nil) {
-//            result = [self appendBodyData:data forKey:currentMultipartBodyKey];
-//        } else if ( currentMultipartFileKey == nil && self.files != nil ) {
-//            result = [self appendFileData:data forKey:currentMultipartFileKey];
-//        }
-//
-//    }
+    // Search for a boundary
+    NSRange searchRange = NSMakeRange(0, data.length);
+    NSRange nextBoundaryRange = [data rangeOfData:self.multipartBoundaryPrefixedData options:0 range:searchRange];
 
+    if ( nextBoundaryRange.location != NSNotFound ) {                                   // We have a boundary
 
+        NSData* CRLFData = [CRConnection CRLFData];
+        NSData* CRLFCRLFData = [CRConnection CRLFCRLFData];
+
+        // Check if we have something before the boundary
+        if ( nextBoundaryRange.location != 0 ) {                                        // There is an existing chunk
+
+            // Extract the piece
+            NSData* preambleData = [NSData dataWithBytesNoCopy:(void *)data.bytes length:nextBoundaryRange.location - CRLFData.length freeWhenDone:NO];
+
+            // Check if we have something buffered
+            if ( self.bufferedBodyData.length > 0 ) {                                   // This is part of a field that
+
+                // Prepend the the buffered data to the one piece
+                NSMutableData* bufferedAndPreambleData = [NSMutableData dataWithCapacity:(self.bufferedBodyData.length + nextBoundaryRange.location)];
+                [bufferedAndPreambleData appendData:self.bufferedBodyData];
+                [bufferedAndPreambleData appendData:preambleData];
+
+                // Flush the buffered data
+                self.bufferedBodyData = nil;
+
+                // Call this method again with the combined data
+                result = [self parseMultipartBodyDataChunk:bufferedAndPreambleData error:error];
+
+            } else {
+
+                // Append the piece to the target if there is one otherwise discard it
+                // (RFC 1341 says to discard anything before the first --boundary (the preamble)
+                // http://www.w3.org/Protocols/rfc1341/7_2_Multipart.html
+
+                if (currentMultipartBodyKey != nil) {
+                    result = [self appendBodyData:preambleData forKey:currentMultipartBodyKey];
+                } else if ( currentMultipartFileKey != nil) {
+                    result = [self appendFileData:preambleData forKey:currentMultipartFileKey];
+                }
+            }
+
+            if ( result ) {
+                // Extract the remaining data
+                NSData* nextChunkData = [NSData dataWithBytesNoCopy:(void *)data.bytes + nextBoundaryRange.location length:data.length - nextBoundaryRange.location freeWhenDone:NO];
+
+                // Call this method again with the remaining data
+                result = [self parseMultipartBodyDataChunk:nextChunkData error:error];
+            }
+
+        } else {                                                                        // We are starting with a new field
+
+            // Read the header (starts after the --boundary and a CRLF and ends with CRLFCRLF)
+            NSUInteger headerStartLocation = nextBoundaryRange.location + nextBoundaryRange.length + CRLFData.length;
+            NSRange headerSearchRange = NSMakeRange(headerStartLocation, data.length - headerStartLocation);
+            NSRange headerTerminatorRange = [data rangeOfData:CRLFCRLFData options:0 range:headerSearchRange];
+
+            if ( headerTerminatorRange.location != NSNotFound ) {                                 // We have a header - all good
+
+                NSData* headerData = [NSData dataWithBytesNoCopy:(void *)data.bytes + headerSearchRange.location length:headerTerminatorRange.location - headerSearchRange.location freeWhenDone:NO];
+                NSString* headerString = [[NSString alloc] initWithBytesNoCopy:(void *)headerData.bytes length:headerData.length encoding:NSUTF8StringEncoding freeWhenDone:NO];
+
+                // Parse the header
+                NSMutableDictionary<NSString *, NSString *> * headerFields = [NSMutableDictionary dictionary];
+                NSArray<NSString *> * headerLines = [headerString componentsSeparatedByString:@"\r\n"];
+                [headerLines enumerateObjectsUsingBlock:^(NSString * _Nonnull headerLine, NSUInteger idx, BOOL * _Nonnull stop) { @autoreleasepool {
+                    NSArray* headerComponents = [headerLine componentsSeparatedByString:CRRequestHeaderNameSeparator];
+                    NSString* headerName = [headerComponents[0] stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceCharacterSet]].lowercaseString;
+                    NSString* headerValue = [headerComponents[1] stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceCharacterSet]];
+                    NSArray* headerValueComponents = [headerValue componentsSeparatedByString:CRRequestHeaderSeparator];
+                    [headerValueComponents enumerateObjectsUsingBlock:^(id  _Nonnull obj, NSUInteger idx, BOOL * _Nonnull stop) { @autoreleasepool {
+                        if ( idx == 0 ) {
+                            headerFields[headerName] = [obj stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceCharacterSet]];
+                            return;
+                        }
+
+                        NSArray* headerValueComponentsParts = [obj componentsSeparatedByString:CRRequestValueSeparator];
+                        NSString* headerValueComponentsPartName = [headerValueComponentsParts[0] stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceCharacterSet]];
+                        NSString* headerValueComponentsPartValue = [headerValueComponentsParts[1] stringByTrimmingCharactersInSet:[NSCharacterSet characterSetWithCharactersInString:@"\" "]];
+                        headerFields[headerValueComponentsPartName] = headerValueComponentsPartValue;
+                    }}];
+                }}];
+
+                // Close any file output streams
+                if ( currentMultipartFileKey != nil ) {
+                    [self.files[currentMultipartFileKey] finishWriting];
+                }
+
+                // Set the target for the value
+                currentMultipartBodyKey = nil;
+                currentMultipartFileKey = nil;
+                if ( headerFields[@"name"].length > 0 ) {
+                    if ( headerFields[@"filename"] && headerFields[@"filename"].length > 0 ) {
+                        currentMultipartFileKey = headerFields[@"name"];
+                    } else if ( !headerFields[@"filename"] ) {
+                        currentMultipartBodyKey = headerFields[@"name"];
+                    }
+                }
+
+                // Create the file
+                if ( currentMultipartFileKey != nil ) {
+                    result = [self setFileHeader:headerFields forKey:currentMultipartFileKey];
+                }
+
+                // Extract the remaining data
+                NSData* nextChunkData = [NSData dataWithBytesNoCopy:(void *)data.bytes + headerTerminatorRange.location + headerTerminatorRange.length length:data.length - headerTerminatorRange.location - headerTerminatorRange.length freeWhenDone:NO];
+
+                // Call this method again with the remaining data
+                result = [self parseMultipartBodyDataChunk:nextChunkData error:error];
+
+            } else {
+                // Checkk if this is the end of the message "--boundary--"
+                NSRange terminatorSearchRange = NSMakeRange(nextBoundaryRange.location + nextBoundaryRange.length, data.length - nextBoundaryRange.location - nextBoundaryRange.length);
+                NSRange terminatorCRLFRange = [data rangeOfData:CRLFData options:0 range:terminatorSearchRange];
+
+                NSData * terminatorData = [NSData dataWithBytesNoCopy:(void *)data.bytes + terminatorSearchRange.location length:data.length - terminatorCRLFRange.location freeWhenDone:NO];
+                if ( [terminatorData isEqualToData:self.multipartBoundaryPrefixData] ) {
+                    // This is the end of the message
+                    result = YES;
+                } else {
+                    // There is no header something is very wrong
+                    *error = [NSError errorWithDomain:CRRequestErrorDomain code:CRRequestErrorMalformedBody userInfo:@{NSLocalizedDescriptionKey: NSLocalizedString(@"Unable to parse multipart data.",)}];
+                    result = NO;
+                }
+            }
+
+        }
+
+    } else {                                                                            // This is just a chunk of something
+
+        // Append the data to the target if there is one otherwise discard it
+        // (RFC 1341 says to discard anything before the first --boundary (the preamble)
+        // http://www.w3.org/Protocols/rfc1341/7_2_Multipart.html
+
+        if ( currentMultipartBodyKey != nil ) {
+            result = [self appendBodyData:data forKey:currentMultipartBodyKey];
+        } else if ( currentMultipartFileKey != nil ) {
+            result = [self appendFileData:data forKey:currentMultipartFileKey];
+        }
+
+    }
     return result;
 }
 
@@ -248,30 +317,42 @@
     }
 }
 
-//- (BOOL)appendBodyData:(NSData *)data forKey:(NSString *)key {
-//    NSLog(@"%s %@ => %lu bytes", __PRETTY_FUNCTION__, key, (unsigned long)data.length);
-//
-//    BOOL result = YES;
-//
-//    if ( _body == nil ) {
-//        _body = [NSMutableDictionary dictionary];
-//    }
-//    NSString* dataString = [[NSString alloc] initWithBytesNoCopy:(void *)data.bytes length:data.length encoding:NSUTF8StringEncoding freeWhenDone:NO];
-//    NSMutableDictionary* body = _body;
-//    if ( body[key] == nil ) {
-//        body[key] = [NSMutableString stringWithString:dataString];
-//    } else {
-//        [((NSMutableString *) body[key]) appendString:dataString];
-//    }
-//
-//    return result;
-//}
+- (BOOL)appendBodyData:(NSData *)data forKey:(NSString *)key {
+    if ( _body == nil ) {
+        _body = [NSMutableDictionary dictionary];
+    }
 
-//- (BOOL)appendFileData:(NSData *)data forKey:(NSString *)key {
-//    NSLog(@"%s %@ => %lu bytes", __PRETTY_FUNCTION__, key, (unsigned long)data.length);
-//
-//    return YES;
-//}
+    NSString* dataString = [[NSString alloc] initWithBytesNoCopy:(void *)data.bytes length:data.length encoding:NSUTF8StringEncoding freeWhenDone:NO];
+    if ( self.body[key] == nil ) {
+        self.body[key] = [NSMutableString stringWithString:dataString];
+    } else {
+        [((NSMutableString *) self.body[key]) appendString:dataString];
+    }
+
+    return YES;
+}
+
+- (BOOL)setFileHeader:(NSDictionary *)headerFields forKey:(NSString *)key {
+    if ( self.files == nil ) {
+        _files = [NSMutableDictionary dictionary];
+    }
+
+    CRUploadedFile * file = [[CRUploadedFile alloc] initWithName:headerFields[@"filename"]];
+    file.mimeType = headerFields[@"content-type"];
+
+    ((NSMutableDictionary *)self.files)[key] = file;
+
+    return YES;
+}
+
+- (BOOL)appendFileData:(NSData *)data forKey:(NSString *)key {
+    if ( self.files[key] == nil ) {
+        return NO;
+    }
+
+    [self.files[key] appendData:data];
+    return YES;
+}
 
 - (void)bufferResponseData:(NSData *)data {
     if ( self.bufferedResponseData == nil ) {
@@ -279,28 +360,6 @@
     } else {
         [self.bufferedResponseData appendData:data];
     }
-}
-
-- (void)validateMultipartBody:(NSData*)bodyData {
-    //    // Validate the input
-    //    NSData* prefixData = [NSData dataWithBytesNoCopy:(void *)bodyData.bytes length:boundaryData.length freeWhenDone:NO];
-    //    if ( [prefixData isEqualToData:boundaryData] ) {
-    //        NSData* suffixData = [NSData dataWithBytesNoCopy:(void *)bodyData.bytes + bodyData.length - boundaryPrefixData.length - 2 length:boundaryPrefixData.length freeWhenDone:NO];
-    //        if ( [suffixData isEqualToData:boundaryPrefixData] ) {
-    //            NSData* lastBoundaryData = [NSData dataWithBytesNoCopy:(void *)bodyData.bytes + bodyData.length - boundaryPrefixData.length - boundaryData.length - 2 length:boundaryData.length freeWhenDone:NO];
-    //            if ( [lastBoundaryData isEqualToData:boundaryData] ) {
-    //            } else {
-    //                result = NO;
-    //                *error = [NSError errorWithDomain:CRRequestErrorDomain code:CRRequestErrorMalformedBody userInfo:@{NSLocalizedDescriptionKey:@"Malformed multipart request.", NSLocalizedFailureReasonErrorKey: [NSString stringWithFormat:@"The body does not end with %@%@", boundaryString, boundaryPrefixString]}];
-    //            }
-    //        } else {
-    //            result = NO;
-    //            *error = [NSError errorWithDomain:CRRequestErrorDomain code:CRRequestErrorMalformedBody userInfo:@{NSLocalizedDescriptionKey:@"Malformed multipart request.", NSLocalizedFailureReasonErrorKey: [NSString stringWithFormat:@"The body does not end with \"%@\"", boundaryPrefixString]}];
-    //        }
-    //    } else {
-    //        result = NO;
-    //        *error = [NSError errorWithDomain:CRRequestErrorDomain code:CRRequestErrorMalformedBody userInfo:@{NSLocalizedDescriptionKey:@"Malformed multipart request.", NSLocalizedFailureReasonErrorKey: [NSString stringWithFormat:@"The body does not start with the boundary. (%@)", boundaryString]}];
-    //    }
 }
 
 - (BOOL)parseURLEncodedBodyData:(NSError *__autoreleasing  _Nullable *)error {
@@ -322,65 +381,61 @@
     return YES;
 }
 
-- (NSString *)description {
-    return [NSString stringWithFormat:@"%@ %@ %@", NSStringFromCRHTTPMethod(self.method), self.URL.path, NSStringFromCRHTTPVersion(self.version)];
+- (NSString *)multipartBoundary {
+    NSString* contentType = _env[@"HTTP_CONTENT_TYPE"];
+    if ([contentType hasPrefix:CRRequestTypeMultipart]) {
+        dispatch_once(&_multipartBoundaryOnceToken, ^{
+            NSArray<NSString*>* headerComponents = [contentType componentsSeparatedByString:CRRequestHeaderSeparator];
+            [headerComponents enumerateObjectsUsingBlock:^(NSString * _Nonnull obj, NSUInteger idx, BOOL * _Nonnull stop) {
+                obj = [obj stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceCharacterSet]];
+                if ( ![obj hasPrefix:CRRequestBoundaryParameter] ) {
+                    return;
+                }
+                _multipartBoundary = [[obj componentsSeparatedByString:CRRequestValueSeparator][1] stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceAndNewlineCharacterSet]];
+            }];
+        });
+    }
+    return _multipartBoundary;
 }
 
-- (BOOL)shouldCloseConnection {
-    BOOL shouldClose = NO;
+- (NSData *)multipartBoundaryPrefixData {
+    static NSData * _multipartBoundaryPrefixData;
+    static dispatch_once_t onceToken;
+    dispatch_once(&onceToken, ^{
+        _multipartBoundaryPrefixData = [NSData dataWithBytesNoCopy:(void * )CRRequestBoundaryPrefix.UTF8String length:CRRequestBoundaryPrefix.length freeWhenDone:NO];
+    });
+    return _multipartBoundaryPrefixData;
+}
 
-    NSString *connectionHeader = [self valueForHTTPHeaderField:@"Connection"];
-    if ( connectionHeader != nil ) {
-        shouldClose = [connectionHeader caseInsensitiveCompare:@"close"] == NSOrderedSame;
-    } else {
-        shouldClose = self.version == CRHTTPVersion1_0;
+- (NSString *)multipartBoundaryPrefixedString {
+    if ( self.multipartBoundary.length > 0 ) {
+        dispatch_once(&_multipartBoundaryPrefixedStringOnceToken, ^{
+            _multipartBoundaryPrefixedString = [NSString stringWithFormat:@"%@%@", CRRequestBoundaryPrefix, self.multipartBoundary];
+        });
+    }
+    return _multipartBoundaryPrefixedString;
+}
+
+- (NSData *)multipartBoundaryPrefixedData {
+    if ( self.multipartBoundaryPrefixedString.length > 0 ) {
+        dispatch_once(&_multipartBoundaryPrefixedDataOnceToken, ^{
+            _multipartBoundaryPrefixedData = [NSData dataWithBytesNoCopy:(void *)self.multipartBoundaryPrefixedString.UTF8String length:self.multipartBoundaryPrefixedString.length freeWhenDone:NO];
+        });
+    }
+    return _multipartBoundaryPrefixedData;
+}
+
+- (void)dealloc {
+    // Delete temporary files created by multipart uploadds
+    if ( self.files.count != 0 ) {
+        [self.files enumerateKeysAndObjectsWithOptions:NSEnumerationConcurrent usingBlock:^(NSString * _Nonnull key, CRUploadedFile * _Nonnull obj, BOOL * _Nonnull stop) {
+            NSURL * temporaryFileURL = obj.temporaryFileURL;
+            dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_BACKGROUND, 0), ^{
+                [[NSFileManager defaultManager] removeItemAtURL:temporaryFileURL error:nil];
+            });
+        }];
     }
 
-    return shouldClose;
 }
-
-//- (NSString *)multipartBoundary {
-//    NSString* contentType = _env[@"HTTP_CONTENT_TYPE"];
-//    if ([contentType hasPrefix:CRRequestTypeMultipart]) {
-//        dispatch_once(&_multipartBoundaryOnceToken, ^{
-//            NSArray<NSString*>* headerComponents = [contentType componentsSeparatedByString:CRRequestHeaderSeparator];
-//            [headerComponents enumerateObjectsUsingBlock:^(NSString * _Nonnull obj, NSUInteger idx, BOOL * _Nonnull stop) {
-//                obj = [obj stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceCharacterSet]];
-//                if ( ![obj hasPrefix:CRRequestBoundaryParameter] ) {
-//                    return;
-//                }
-//                _multipartBoundary = [[obj componentsSeparatedByString:CRRequestValueSeparator][1] stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceAndNewlineCharacterSet]];
-//            }];
-//        });
-//    }
-//    return _multipartBoundary;
-//}
-//
-//- (NSData *)multipartBoundaryPrefixData {
-//    static NSData * _multipartBoundaryPrefixData;
-//    static dispatch_once_t onceToken;
-//    dispatch_once(&onceToken, ^{
-//        _multipartBoundaryPrefixData = [NSData dataWithBytesNoCopy:(void * )CRRequestBoundaryPrefix.UTF8String length:CRRequestBoundaryPrefix.length freeWhenDone:NO];
-//    });
-//    return _multipartBoundaryPrefixData;
-//}
-//
-//- (NSString *)multipartBoundaryPrefixedString {
-//    if ( self.multipartBoundary.length > 0 ) {
-//        dispatch_once(&_multipartBoundaryPrefixedStringOnceToken, ^{
-//            _multipartBoundaryPrefixedString = [NSString stringWithFormat:@"\r\n%@%@", CRRequestBoundaryPrefix, self.multipartBoundary];
-//        });
-//    }
-//    return _multipartBoundaryPrefixedString;
-//}
-//
-//- (NSData *)multipartBoundaryPrefixedData {
-//    if ( self.multipartBoundaryPrefixedString.length > 0 ) {
-//        dispatch_once(&_multipartBoundaryPrefixedDataOnceToken, ^{
-//            _multipartBoundaryPrefixedData = [NSData dataWithBytesNoCopy:(void *)self.multipartBoundaryPrefixedString.UTF8String length:self.multipartBoundaryPrefixedString.length freeWhenDone:NO];
-//        });
-//    }
-//    return _multipartBoundaryPrefixedData;
-//}
 
 @end