+++ Criollo/Source/CRRequest_Internal.h
@@ -8,6 +8,7 @@
 
 #import "CRRequest.h"
 
+#define CRRequestHeaderNameSeparator        @":"
 #define CRRequestHeaderSeparator            @";"
 #define CRRequestHeaderArraySeparator       @","
 #define CRRequestKeySeparator               @"&"
@@ -16,6 +17,7 @@
 #define CRRequestBoundaryPrefix             @"--"
 
 NS_ASSUME_NONNULL_BEGIN
+
 @interface CRRequest ()
 
 @property (nonatomic, readonly) BOOL shouldCloseConnection;
@@ -36,8 +38,9 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)bufferBodyData:(NSData *)data;
 - (void)bufferResponseData:(NSData *)data;
 
-//- (BOOL)appendBodyData:(NSData *)data forKey:(NSString *)key;
-//- (BOOL)appendFileData:(NSData *)data forKey:(NSString *)key;
+- (BOOL)appendBodyData:(NSData *)data forKey:(NSString *)key;
+- (BOOL)setFileHeader:(NSDictionary *)headerFields forKey:(NSString *)key;
+- (BOOL)appendFileData:(NSData *)data forKey:(NSString *)key;
 
 - (void)setEnv:(NSDictionary<NSString *,NSString *> *)envDictionary;
 - (void)setEnv:(NSString *)obj forKey:(NSString *)key;
@@ -50,4 +53,5 @@ NS_ASSUME_NONNULL_BEGIN
 - (BOOL)parseBufferedBodyData:(NSError * _Nullable __autoreleasing * _Nullable)error;
 
 @end
+
 NS_ASSUME_NONNULL_END